+++ test/test_report.py
@@ -33,9 +33,9 @@ def parse_table(table):
     for row in table.find("tbody").find_all("tr"):
         d = {k: v.get_text() for k, v in zip(headers, row.find_all("td"))}
         # Convert the price field to float
-        for k, v in d.items():
-            if "€" in v:
-                d[k] = price_to_float(v)
+        # for k, v in d.items():
+        #     if "€" in v:
+        #         d[k] = price_to_float(v)
         data.append(d)
     return data
 
@@ -65,6 +65,21 @@ def test_variant_overview(workflow_dir):
 
     assert gene2["rowspan"] == "1"
     assert gene2.text == "MT-ATP8"
+
+
+
+@pytest.mark.workflow('test-report')
+def test_genes_in_order(workflow_dir):
+    """Test if genes in table 2 are in order"""
+    report = f"{workflow_dir}/report.html"
+    with open(report) as fin:
+        soup = bs4.BeautifulSoup(fin, features="html.parser")
+
+    # Extract table 2, genes of interest
+    genes_of_interest = parse_table(soup.find('table', id='genes-of-interest'))
+    assert "MT-ATP6" in genes_of_interest[0].values()
+    assert "MT-ATP8" in genes_of_interest[1].values()
+
 
 
 @pytest.mark.workflow('test-report')